requirements.txt
GitPython
PyYaml
